-74HC595 SLİDİNG LED
-74HC595 SLİDİNG LED
-74HC595 SLİDİNG LED

The 74HC595 Shift register IC is a shift register that provides parallel reception of serial input data. By storing the serial data in D type flip-flop, it provides parallel output in the desired state. 74HC595 ICs can be connected to each other in a cascade manner thanks to its 9th pin. It is generally used in light show applications. It can be easily controlled by Arduino or any other microprocessor.

 

• Bit: 8

• Pin: 16

• Logic family: HC

• Logical function: Shift register

• Propagation delay time: 285ns

• Operating supply voltage (Type): 2.5V/3.3V/5V

• Operating temperature: -55 °C ~ 125 °C

• Operating supply voltage (Min): 2V

• Operating supply voltage (Max): 6V

• Installation: Through Hole

• Package: 16-DIP (0.300"/7.62mm)

 





-74HC595

- 220R RESİSTOR X8

-RED LED X4

-GREEN LED X4

-SWİTCH X4

-10K RESİSTOR X3

-1K RESİSTOR X2

-0.1 uF CAPASİTOR